Remove visibility handling storage query factory interface and simplify code near that.
