--- rdbms/src/main/java/org/spine3/server/storage/jdbc/entity/visibility/AbstractVisibilityHandlingStorageQueryFactory.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.spine3.server.storage.jdbc.entity.visibility;
-
-import org.slf4j.Logger;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
-/**
- * An implementation base for {@link VisibilityHandlingStorageQueryFactory}.
- *
- * @author Dmytro Dashenkov.
- */
-public abstract class AbstractVisibilityHandlingStorageQueryFactory<I>
-        implements VisibilityHandlingStorageQueryFactory<I>{
-
-    private Logger logger;
-
-    /**
-     * Sets the {@link Logger} to use in the generated queries.
-     */
-    public void setLogger(Logger logger) {
-        this.logger = checkNotNull(logger);
-    }
-
-    protected Logger getLogger() {
-        return logger;
-    }
-}
